# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: VulnOps Hub
  description: Vulnerability-operations control plane — modular monolith
  version: 0.1.0
paths:
...
  /api/v1/organizations/{org_id}/services:
    post:
      tags:
      - services
      summary: Create Business Service
      operationId: create_business_service_api_v1_organizations__org_id__services_post
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessServiceResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              additionalProperties: false
              description: Organization-scoped business-service inventory write contract.
              properties:
                name:
                  minLength: 1
                  title: Name
                  type: string
                owner_team:
                  minLength: 1
                  title: Owner Team
                  type: string
                criticality:
                  anyOf:
                  - minLength: 1
                    type: string
                  - type: 'null'
                  title: Criticality
              required:
              - name
              - owner_team
              title: BusinessServiceCreateRequest
              type: object
    get:
      tags:
      - services
      summary: List Business Services
      operationId: list_business_services_api_v1_organizations__org_id__services_get
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessServiceListResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/organizations/{org_id}/services/{service_id}:
    get:
      tags:
      - services
      summary: Get Business Service
      operationId: get_business_service_api_v1_organizations__org_id__services__service_id__get
      security:
      - BearerAuth: []
      parameters:
      - name: org_id
        in: path
        required: true
        schema:
          type: string
          title: Org Id
      - name: service_id
        in: path
        required: true
        schema:
          type: string
          title: Service Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessServiceResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BusinessServiceListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/BusinessServiceResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - total
      title: BusinessServiceListResponse
    BusinessServiceResponse:
      properties:
        id:
          type: string
          title: Id
        organization_id:
          type: string
          title: Organization Id
        name:
          type: string
          title: Name
        owner_team:
          type: string
          title: Owner Team
        criticality:
          anyOf:
          - type: string
          - type: 'null'
          title: Criticality
        created_at:
          type: string
          title: Created At
        updated_at:
          type: string
          title: Updated At
      type: object
      required:
      - id
      - organization_id
      - name
      - owner_team
      - created_at
      - updated_at
      title: BusinessServiceResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ProblemDetails:
      properties:
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        status:
          type: integer
          title: Status
        code:
          type: string
          title: Code
        detail:
          type: string
          title: Detail
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        fields:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Fields
      type: object
      required:
      - type
      - title
      - status
      - code
      - detail
      title: ProblemDetails
      description: Stable error envelope shared by authenticated API operations.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    BearerAuth:
      type: http
      description: OIDC access token
      scheme: bearer
      bearerFormat: JWT